MAINT: move buffer.h -> npy_buffer.h to avoid conflicts
