+++ numpy/core/src/multiarray/arraytypes.c.src
@@ -36,7 +36,7 @@
 
 #include "cblasfuncs.h"
 #include "npy_cblas.h"
-#include "buffer.h"
+#include "npy_buffer.h"
 
 /* check for sequences, but ignore the types numpy considers scalars */
 static NPY_INLINE npy_bool
